fixed location of working dir for macro
(cherry picked from commit 541c0153cc6077669505cd050c57ff60b3b74083)
(cherry picked from commit 4e9ad75970d148e77e7f350e069136dfb9591a96)
(cherry picked from commit 10a467dfbd35f9d3a4b820d347642915a8f375cd)
(cherry picked from commit 20fbcd6cbf3bf6d218449f57ee1a0c6af8c698ba)

+++ private/benchmarks/benchmark-infra/src/main/worker/bootstrap-worker.sh
@@ -25,14 +25,14 @@ done
 cd /work
 
 worker_artifact=/work/benchmark-worker.jar
-work_dir=$(pwd)/macro_work_dir/
+work_dir=/work/run/
 
-rm -rf "${worker_artifact}" "${work_dir}"
+# make sure we start clean with working directory
+# and worker artifact
+rm -rf "${worker_artifact}" "${work_dir:?}/*"
 
 # download bootstrap jar
 aws --region eu-north-1 s3 cp "${workerArtifactUri}" "${worker_artifact}"
-
-mkdir "${work_dir}"
 
 # shellcheck disable=SC2086
 java ${JAVA_OPTS:+"$JAVA_OPTS"} -jar benchmark-worker.jar "${params[@]}" --workspace-dir "${work_dir}" --batch-job-id "${AWS_BATCH_JOB_ID}"
